refactor(ast): import extractStringsFromNodes from its own file, not the utils barrel
Extract extractStringsFromNodes into utils/extractStringsFromNodes.ts so the
root barrel (src/index.ts) and factory.ts import it directly from its
definition file instead of pulling the @kubb/ast/utils barrel. utils/index.ts
re-exports it, so the @kubb/ast/utils subpath is unchanged. No public API change.
